Add InteractiveNotice.
